# webapp/src/main/resources/docs/API.yaml
info:
  title: API Documentation
  version: 1.0
openapi: 3.0.0
paths:
  /appointment:
    get:
      summary: Get All Appointments
      responses:
        '200':
          description: OK
    post:
      summary: Create new UserTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                patient:
                  type: object
                  properties:
                    id:
                      type: integer
                dentist:
                  type: object
                  properties:
                    id:
                      type: integer
      responses:
        '201':
          description: Created
  /appointment/{id}:
    get:
      summary: Get UserTask By ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    put:
      summary: Update UserTask By ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
      responses:
        '200':
          description: OK
    delete:
      summary: Delete UserTask By ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
  /dentist:
    get:
      summary: Get All Dentists
      responses:
        '200':
          description: OK
    post:
      summary: Create new Comment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                surname:
                  type: string
                licenseNumber:
                  type: integer
      responses:
        '201':
          description: Created
  /dentist/{id}:
    get:
      summary: Get Comment By ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    put:
      summary: Update Comment By ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                surname:
                  type: string
      responses:
        '200':
          description: OK
    delete:
      summary: Delete Comment By ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
  /patient:
    get:
      summary: Get All Patients
      responses:
        '200':
          description: OK
    post:
      summary: Create new Patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                surname:
                  type: string
                registrationDate:
                  type: string
                niNumber:
                  type: integer
                address:
                  type: object
                  properties:
                    streetName:
                      type: string
                    streetNumber:
                      type: integer
                    floor:
                      type: integer
                    department:
                      type: string
      responses:
        '201':
          description: Created
  /patient/{id}:
    get:
      summary: Get Patient By ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    put:
      summary: Update Patient By ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                surname:
                  type: string
                address:
                  type: object
                  properties:
                    streetName:
                      type: string
                    streetNumber:
                      type: integer
                    floor:
                      type: integer
                    department:
                      type: string
      responses:
        '200':
          description: OK
  /user:
    get:
      summary: Get All Users
      responses:
        '200':
          description: OK
    post:
      summary: Create new User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                role:
                  type: string
      responses:
        '201':
          description: Created
  /user/{id}:
    get:
      summary: Get User By ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    put:
      summary: Update User By ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                role:
                  type: string
      responses:
        '200':
          description: OK
    delete:
      summary: Delete User By ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content